Refactor out ai requests into own router
